backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.8
openai==1.3.7
faiss-cpu==1.7.4
numpy==1.24.3
python-dotenv==1.0.0
pydantic==2.5.0
